[BUGFIX] Avoid undefined array key `showitem`
It is recommended to extending global user settings
within the extension `ext_tables.php` file looking
into the documentation [1].

Mostly correct, adding the custom user setting field
is done manually not really taking care if the showitem
array key already exists or not.

TYPO3 provides a safe way for extension authors to do that
with `ExtensionManagementUtility::addFieldsToUserSettings`
and uses this now to avoid the

  1) /var/www/html/ext_tables.php:16
  Undefined array key "showitem"

  Triggered by:

  * EHAERER\PasteReference\Tests\Unit\DummyTest::extensionLoaded
    /var/www/html/Tests/Functional/DummyTest.php:40

error revealed by the recently added dummy functional test.

That method would allow to easily controll adding the custom
setting before or after other settings, but is left out for
now to keep the current setting.

[1] https://docs.typo3.org/m/typo3/reference-coreapi/12.4/en-us/Configuration/UserSettingsConfiguration/Extending.html