[TASK] Introduce `Build/Scripts/runTests.sh` dispatcher
TYPO3 Core provides a conveniant toolchain dispatcher
script called `Build/Scripts/runTests.sh` allowing to
execute commands in controlled environments using the
container technology (docker or podman). Quite a couple
extension authors adopted in the past this script to
provide a equialent experience when working on their
extension or ease the process for contributors.

Main point for this script is, that it can easily be
executed in different CI environment and host systems
as long as docker or podman as containerization tools
are available.

This change now introduces a generic adoption of the
full-packed TYPO3 Core Development template with first
execution abilities:

* `-s composer` command dispatcher as alternative for
  `ddev composer` if someone do not like to use ddev.
* `-s unit` to execute unit tests
* `-s functional` to execute functional test including
  the abilitiy for multi php-version, database driver,
  database vendor and database vendor version selection.

For available desription see the help text provided by
the script using

```bash
Build/Scripts/runTests.sh
```

Note that we do not use the core file directly and use
a stripped down variant to reduce required adoption to
a bare minimum for now.

Used commands(s):

```bash
mkdir -p Build/Scripts \
  && TMP_PREFIX="https://raw.githubusercontent.com/" \
  && TMP_REPOSITORY="sbuerk/tf-basics-extension" \
  && TMP_GITREF="/refs/heads/main/.demo-templates/" \
  && TMP_PATH="Build/Scripts/runTests.sh" \
  && TMP_URL="${TMP_PREFIX}${TMP_REPOSITORY}" \
  && TMP_URL="${TMP_URL}${TMP_GITREF}${TMP_PATH}" \
  && curl -L -o Build/Scripts/runTests.sh ${TMP_URL} \
  && chmod a+x Build/Scripts/runTests.sh \
  && ddev exec \
    "sed -i 's/.Build\/Web\//.Build\/public\//g' Build/Scripts/runTests.sh" \
  && ddev exec \
    'sed -i -e "s/\${ROOT_DIR}\/public\//\${ROOT_DIR}\/.Build\/public\//g" Build/Scripts/runTests.sh'
```