[TASK] Replace `\PDO::*` constants with `Connection::PARAM_*`
In old days using `\PDO::` constants for database
parameter type declarations was a thing, and even
Doctrine DBAL supported the `php-ext:PDO` constants.

Since Doctrine DBAL v3 the usage of these constants
are discouraged and the recommended migration path
is to use the ArrayParameterType or ParameterType
class constans directly, which will be replaced by
native PHP enum implementation in Doctrine DBAL v4
upcoming in TYPO3 v13.

The TYPO3 Core Team provided on the decorated DBAL
connection implementation class constants mapping
to the \PDO constants and replaces them in TYPO3 v13
with the new Doctrine DBAL php enums and recommend
to use the `Connection::PARAM_*` constants directly
to be prepared for an upgraded, which is now done
with this change.

That allows dropping the direct dependency to the
`php-ext:PDO` in the extension `composer.json`.

Used command(s):

```bash
ddev composer remove "ext-pdo"
```