[TASK] Apply `php-cs-fixer` changes
Recently the `php-cs-fixer` has been added as
a tool to ensure coding style guidelines along
with a rule-set.

However, the already existing code does not match
this rule-set and the detected changes are added
now to have a clean base-line.

Note that this is a pre-requisit to enable later
php code cgl checks in automatic GitHub action
pipelines to ensure cgl for contribution and
work prior to merging code.

Used command(s):

```bash
Build/Scripts/runTests.sh -s cgl
```