[TASK] Prepare new file-header as `php-cs-fixer` rule
This change prepares the rule definition to ensure
PHP file header confirmity using the recently added
`php-cs-fixer` binary.

The rule is commented-out for now to discuss the correct
license and header to use and apply to all files in a
later step.

Just acting as a example / test configurtion which can
be easily activated.